Adjust int volumes I missed

+++ src/codecs/native_midi/native_midi_macosx.c
@@ -396,7 +396,7 @@ bool native_midi_active(void)
 {
 }
 
-void native_midi_setvolume(int volume)
+void native_midi_setvolume(float volume)
 {
 }
